caption_api/requirements.txt
#ImageMagick instalação -> https://www.tecmint.com/install-imagemagick-on-debian-ubuntu/  + executar sudo apt install libmagick++-dev
# Verificar dpkg -l | grep magic

moviepy
flask
flask-restful
awscli
boto3
flask-cors
python-dotenv